K8S-2108: Refactor and unit tests
Refactored to reduce complexity and also removed the disabling of this lint check to confirm.
Refactored into smaller library code components
Added unit tests as well to ensure no breakage as part of this and improve the overall product.
Moved into sub-directory following package-layout recommendations to also allow for linting on build machine.
Switched to Uber Zap logging library to align with operator.
Using config structs as well instead of random strings which are easier to order wrong.

Change-Id: If2c2718eadc47ebb47abac2bcf6e99f530d4cab9
Reviewed-on: http://review.couchbase.org/c/couchbase-fluent-bit/+/150856
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>