K8S-2108: Reduce refresh interval
Testing with various platforms has demonstrated that it might be useful to reduce the refresh interval when checking for logs.
Even though the container starts very fast, it has to wait for the Couchbase Server to create the directories and then may miss logs for up to 60s (default) before checking again.

Change-Id: Ic1432093558136787b4be9ec1eb4d04f8f42d14e
Reviewed-on: http://review.couchbase.org/c/couchbase-fluent-bit/+/151522
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>