K8S-2108: Add information for release tagging and branching
Added a quick statement about the expected approach for tagging and branching.
I cannot find other examples of this in Couchbase documentation at the moment.
Ultimately this is industry standard approach: hopefully we can get the release process to add a tag automatically.

Also added a note on the multiline output.

Change-Id: I288d3078d0c30289c9f098b36a579d9a532587bb
Reviewed-on: http://review.couchbase.org/c/couchbase-fluent-bit/+/151122
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Eric Schneider <eric.schneider@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>