K8S-2159: Support non-root running
Updated to follow same approach as operator to ensure we run as non-root, plus a test to run as uid 1000.
This found a few issues in the test script as well.
Identified that not all service labels are present which triggers some failures in testing as well so removed those too.
Added helper script to flatten config too.
Added shellcheck as well for scripts available, on top of existing Hadolint for container definitions.

Change-Id: I64011d8ecad3ba168f3a90a8871e427dd4ab99da
Reviewed-on: http://review.couchbase.org/c/couchbase-fluent-bit/+/152558
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>