K8S-2147: Resolve issues with 'multi-format' logs creating wrong levels
The level is corrected to UNKNOWN if it is not parsed properly.
The original level is maintained for reference, it is not easy to append it to the message field without invoking LUA scripting.
LogQL or similar could do this anyway though downstream - all the info is there and kept simple.

Added filename shortening to make it easier to use with Grafana as well.

Change-Id: Ia1cf920914d806336faffc3a99bcd66cd9637272
Reviewed-on: http://review.couchbase.org/c/couchbase-fluent-bit/+/157578
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>