K8S-2105: Server log verification
Aimed at verifying the sidecar with the step up to 7, this change also adds an option to run general logging tests with any server image we support.
We override the entrypoint to allow us to have declarative control.

This will run a test case that uses the Fluent Bit EXPECT plugin:
Every log record we handle will be checked to see if it contains all the expected fields (it won't if it is not parsed).
If any are missing then it will exit.
We then monitor the number of containers running to ensure both stay up until the end.

As part of this we have also added support for the new CBS 7 Prometheus log and some Red Hat tweaks to resolve CVEs.

Change-Id: Ie6f6af174f1ec5326e42de0de37ba2f11cafb8df
Reviewed-on: http://review.couchbase.org/c/couchbase-fluent-bit/+/157872
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>