K8S-2296: Resolve error in latest config
Cunningly there is a typo in the include declaration for the new Prometheus log file.
Helpfully Fluent Bit just gives a generic warning about it and it only shows up when using the full configuration all together rather than testing each input separately.
A couple of extra checks have therefore been added to verify the configuration during build and test to try to prevent this in the future.

Debugging this issue showed some deficiencies in logging which have also been resolved as well as ensuring the integration test does actually fail when it should.

Change-Id: I72641ec9c2c26c3c9cabe0e32bd6588cb9e1e358
Reviewed-on: http://review.couchbase.org/c/couchbase-fluent-bit/+/158230
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>